feat: small changes to files for ts
