+++ src/index.ts
@@ -1,14 +1,14 @@
 import removeDecorators from './parse/removeDecorators.js';
 import redecorate from './parse/redecorate.js';
-import compile from './solc.mjs';
+import compile from './solc.js';
 
-// import checks from './transformers/checks.js';
+import checks from './transformers/checks.js';
 // import ownership from './transformers/ownership.mjs';
 // import toCircuit from './transformers/toCircuit.mjs';
 // import toContract from './transformers/toContract.mjs';
 // import toOrchestration from './transformers/toOrchestration.mjs';
 
-const zappify = options => {
+const zappify = (options: any) => {
   // Error.stackTraceLimit = 0; // prettier error output TODO see if we can remove this in place of error/errors.mjs handling the limit
   // testing prettier errors
   process.on('uncaughtException', err => {
@@ -22,7 +22,7 @@ const zappify = options => {
 
   const zolAST = redecorate(solAST, toRedecorate, options);
 
-  // let path = checks(zolAST, options);
+  let path = checks(zolAST);
   //
   // path = ownership(path, options);
   //
@@ -35,7 +35,7 @@ const zappify = options => {
   // toContract(zolAST, options);
   //
   // if (options.isTest) return path.scope.indicators;
-  console.log(zolAST);
+  console.log(path.node.nodes[1]);
   return zolAST;
 };
